Moved a commonly used configuration file to source folder and renamed
