+++ Source/zx81/zx81.h
@@ -22,7 +22,7 @@
 #ifndef zx81_h
 #define zx81_h
 
-#include "z80\config.h"
+#include "configuration.h"
 
 extern BYTE memory[];
 extern BYTE acecolour[];
